Generate and return user tokens upon login
This commit adds a new dependency: jsonwebtoken, which is used to
generate the tokens according to RFC 7519 - JSON Web Tokens.

Because there are now multiple relevant pieces of data about a user, the
route /role has been replaced by the route /userdata, which accepts a
token and returns all important pieces of data about a user, including
the username.

This commit's purpose is to improve security, see discussion on the
corresponding frontend commit.